Added Revolve Display and Sensor Programmed Logic
This is the main code used for my senior project "Revolve". Its an interactive display used to save/show user account data collected from various sensors, and store that data to the currently logged in account. Hosted in the Arduino IDE and written using C++, it does require the usage of specific sensors/components to run, so please take a look at the readme if you decide to recreate this system!