Added Touchscreen Communication Interface Files
Included are the touchscreen interface files (later repurposed to be used in Revolve) its goal was to be able to both read/send user data between a connected computer using serial communication and the spi interface to do so. This project does use libraries exclusive to the ILI9341 Display i used so please be warned another display might have a different implementation in regards to the dimensions/drivers.